Remove support for building a shared library
Because uSTL does not use the entire contents of libsupc++, allowing a
shared library results in having to link -lsupc++ in anyway, thus removing
any convenience of the shared library as a unit and giving missing
symbols errors on link, which can be baffling to the inexperienced. So,
from now on, only the static library is built and installed. Apps using
the library should link with gcc and `pkg-config --libs ustl`.